# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Templet Webinars API",
        "description": "API de solo lectura para consultar submissions, asistentes y estadísticas de los webinars. Autenticación por API key en el header `X-Api-Key`.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Servidor actual"
        }
    ],
    "paths": {
        "/api/v1/attendees": {
            "get": {
                "tags": [
                    "Attendees"
                ],
                "summary": "Lista los asistentes reales (Zoom) de los webinars",
                "description": "Participantes sincronizados desde los reportes de Zoom mediante `php artisan webinars:sync-attendees`. Incluye hora de entrada/salida y duración.",
                "operationId": "listAttendees",
                "parameters": [
                    {
                        "name": "client",
                        "in": "query",
                        "description": "Slug o ID del cliente",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "libertynet"
                        }
                    },
                    {
                        "name": "webinar_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "webinar_slug",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "campaign",
                        "in": "query",
                        "description": "Campaign Salesforce Field del webinar",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "join_time desde (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "join_time hasta (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "min_duration",
                        "in": "query",
                        "description": "Duración mínima en minutos",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 50
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado paginado de asistentes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Attendee"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "API key faltante o inválida"
                    },
                    "403": {
                        "description": "La API key no tiene acceso a ese cliente"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Attendee": {
                "title": "Attendee",
                "description": "Asistente real al webinar, sincronizado desde los reportes de Zoom.",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 88
                    },
                    "webinar": {
                        "$ref": "#/components/schemas/WebinarRef"
                    },
                    "name": {
                        "type": "string",
                        "example": "Ana Pérez",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "example": "ana@empresa.com",
                        "nullable": true
                    },
                    "join_time": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "leave_time": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "duration": {
                        "description": "Duración en segundos",
                        "type": "integer",
                        "example": 2700
                    },
                    "duration_minutes": {
                        "type": "integer",
                        "example": 45
                    },
                    "device": {
                        "type": "string",
                        "example": "Windows",
                        "nullable": true
                    },
                    "location": {
                        "type": "string",
                        "example": "Mexico City, Mexico",
                        "nullable": true
                    },
                    "zoom_participant_id": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "WebinarRef": {
                "title": "WebinarRef",
                "description": "Referencia al webinar y su cliente.",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 12
                    },
                    "title": {
                        "type": "string",
                        "example": "LibertyNet SASE"
                    },
                    "slug": {
                        "type": "string",
                        "example": "libertynet-sase"
                    },
                    "campaign_salesforce_field": {
                        "description": "Campo `campaign` del webinar. Es el identificador con el que se enlaza la campaña en Salesforce.",
                        "type": "string",
                        "example": "7013h000000abcAAA",
                        "nullable": true
                    },
                    "client": {
                        "$ref": "#/components/schemas/ClientRef"
                    }
                },
                "type": "object"
            },
            "ClientRef": {
                "title": "ClientRef",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 3
                    },
                    "name": {
                        "type": "string",
                        "example": "LibertyNet"
                    },
                    "slug": {
                        "type": "string",
                        "example": "libertynet"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "API key generada con `php artisan api-key:create`.",
                "name": "X-Api-Key",
                "in": "header"
            }
        }
    },
...
}